+++ frontend/cfb/resources/team.js
