+++ frontend/static/schedule.json
